refactor: disassemble the web UI into different packages
I have a real reason for doing this: The backend relies on signals, but solid's
signals don't trigger effects on the server side, so I need to replace the
signal implementation for the server side codes. Extracting the server side
code into a separate package will make it clear which part of the codebase uses
which signal implementation.